Move llava next functions to pipeline after merge with changed vlm structure
